+++ packages/riverpod/test/old/framework/provider_element_test.dart
@@ -4,7 +4,7 @@ import 'package:mockito/mockito.dart';
 import 'package:riverpod/src/internals.dart';
 import 'package:test/test.dart';
 
-import '../utils.dart';
+import '../../utils.dart';
 
 void main() {
   group('Ref.exists', () {
