+++ packages/riverpod/test/old/framework/ref_test.dart
@@ -2,7 +2,7 @@ import 'package:mockito/mockito.dart';
 import 'package:riverpod/riverpod.dart';
 import 'package:test/test.dart';
 
-import '../utils.dart';
+import '../../utils.dart';
 
 void main() {
   group('Ref', () {
